Move AppComponentsFactory to its own file
